feat: Improve error handling in deleteProject method

+++ src/shared/infra/repositories/project_repository_prisma.ts
@@ -55,9 +55,6 @@ export class ProjectRepositoryPrisma implements IProjectRepository {
 
   async getProjectById(projectId: string): Promise<Project | undefined> {
     try {
-      console.log("ENTRANDO NO REPO DO PROJETO")
-      console.log("Buscando projeto no banco de dados...");
-      console.log("ID do projeto:", projectId)
       const projectFromPrisma = await prisma.project.findUnique({
         where: {
           project_id: projectId,
@@ -109,9 +106,19 @@ export class ProjectRepositoryPrisma implements IProjectRepository {
     }
   }
 
-  async deleteProject(projectId: string): Promise<void> { 
+  async deleteProject(projectId: string): Promise<void> {
     try {
-      const project = await prisma.project.delete({
+      const project = await prisma.project.findUnique({
+        where: {
+          project_id: projectId,
+        },
+      });
+  
+      if (!project) {
+        throw new Error("Projeto não encontrado.");
+      }
+  
+      await prisma.project.delete({
         where: {
           project_id: projectId,
         },
@@ -120,5 +127,5 @@ export class ProjectRepositoryPrisma implements IProjectRepository {
       console.error("Erro ao deletar projeto:", error);
       throw new Error("Erro ao deletar projeto no banco de dados.");
     }
-  }
+  }  
 }
